SFML_Tilemapper
This is a library that is a work in progress. It is essentially a library that is designed to parse TMX and TSX files from a map maker program named Tiled. Even though it is labeled to use with SFML, the main goal is to correctly parse TMX in the least headache inducing way possible. Such examples include being able to use the TMX file to automatically deduce encyrption use and be able to correctly parse all TMX files automatically. Of course, this library is completely dependent upon the version of Tiled that it is deisgned for.